fix: move methods of ExchangeRateTriggerService
