+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/service/finance/FinanceTriggerService.java
@@ -2,14 +2,18 @@ package com.ssafy.shinhanflow.service.finance;
 
 import org.springframework.stereotype.Service;
 
+import com.ssafy.shinhanflow.config.error.ErrorCode;
+import com.ssafy.shinhanflow.config.error.exception.BadRequestException;
+import com.ssafy.shinhanflow.dto.finance.exchange.ExchangeRateRequestDto;
+import com.ssafy.shinhanflow.dto.finance.exchange.ExchangeRateResponseDto;
 import com.ssafy.shinhanflow.dto.finance.header.RequestHeaderDto;
 import com.ssafy.shinhanflow.dto.finance.product.DepositAndSavingProductsRequestDto;
 import com.ssafy.shinhanflow.dto.finance.product.DepositAndSavingProductsResponseDto;
 import com.ssafy.shinhanflow.dto.finance.product.LoanProductsRequestDto;
 import com.ssafy.shinhanflow.dto.finance.product.LoanProductsResponseDto;
-import com.ssafy.shinhanflow.service.TriggerService;
 import com.ssafy.shinhanflow.util.FinanceApiHeaderGenerator;
 import com.ssafy.shinhanflow.util.FinanceApiService;
+import com.ssafy.shinhanflow.util.constants.Currency;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,8 +21,8 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class FinanceTriggerService extends TriggerService {
-	private final FinanceApiService financeApiFetcher;
+public class FinanceTriggerService {
+	private final FinanceApiService financeApiService;
 	private final FinanceApiHeaderGenerator financeApiHeaderGenerator;
 
 	public DepositAndSavingProductsResponseDto depositProductsInfo() {
@@ -28,7 +32,7 @@ public class FinanceTriggerService extends TriggerService {
 		DepositAndSavingProductsRequestDto dto = DepositAndSavingProductsRequestDto.builder()
 			.header(header)
 			.build();
-		return financeApiFetcher.depositProductsInfo(dto);
+		return financeApiService.depositProductsInfo(dto);
 	}
 
 	public DepositAndSavingProductsResponseDto savingProductsInfo() {
@@ -38,7 +42,7 @@ public class FinanceTriggerService extends TriggerService {
 		DepositAndSavingProductsRequestDto dto = DepositAndSavingProductsRequestDto.builder()
 			.header(header)
 			.build();
-		return financeApiFetcher.savingProductsInfo(dto);
+		return financeApiService.savingProductsInfo(dto);
 	}
 
 	public LoanProductsResponseDto loanProductsInfo() {
@@ -48,6 +52,22 @@ public class FinanceTriggerService extends TriggerService {
 		LoanProductsRequestDto dto = LoanProductsRequestDto.builder()
 			.header(header)
 			.build();
-		return financeApiFetcher.loanProductsInfo(dto);
+		return financeApiService.loanProductsInfo(dto);
+	}
+
+	public ExchangeRateResponseDto getExchangeRate(String currencyCode) {
+		log.info("getExchangeRate - currencyCode: {}", currencyCode);
+		if (currencyCode == null) {
+			throw new BadRequestException(ErrorCode.NULL_REQUIRED_VALUE);
+		}
+
+		// userKey 필요 없음
+		RequestHeaderDto header = financeApiHeaderGenerator.createHeader("exchangeRate", null, "00100");
+		ExchangeRateRequestDto dto = ExchangeRateRequestDto.builder()
+			.header(header)
+			.currency(Currency.valueOf(currencyCode))
+			.build();
+
+		return financeApiService.getExchangeRate(dto);
 	}
 }
